reconfigured entities structure
