+++ backend/src/main/java/swp391/com/backend/pojo/blog/Blog.java
@@ -1,9 +1,10 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.blog;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import swp391.com.backend.pojo.roles.Account;
 
 import java.time.LocalDateTime;
 import java.util.List;
