+++ backend/src/main/java/swp391/com/backend/pojo/blog/Tag.java
@@ -1,26 +1,31 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.blog;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
-public class PaymentMethod {
+public class Tag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @Column(name = "method_name")
-    private String methodName;
+    @Column(name = "tag_name", unique = true)
+    private String tagName;
 
     @Lob
     private String description;
 
     @Column(name = "is_active")
     private Boolean isActive;
+
+    @OneToMany(mappedBy = "tag")
+    private List<BlogTag> blogTags;
 }
