+++ backend/src/main/java/swp391/com/backend/pojo/order/ResultDetail.java
@@ -1,9 +1,10 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.order;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import swp391.com.backend.pojo.test.Test;
 
 @Entity
 @Getter
