+++ backend/src/main/java/swp391/com/backend/pojo/roles/Account.java
@@ -1,10 +1,11 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.roles;
 
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import swp391.com.backend.pojo.blog.Blog;
 
 import java.util.List;
