+++ backend/src/main/java/swp391/com/backend/pojo/roles/Admin.java
@@ -1,26 +1,21 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.roles;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
-public class Result {
+public class Admin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
     @OneToOne
-    @JoinColumn(name = "order_id", unique = true)
-    private Order order;
-
-    @OneToMany(mappedBy = "result")
-    private List<ResultDetail> resultDetails;
+    @JoinColumn(name = "account_id", unique = true)
+    private Account account;
 }
