+++ backend/src/main/java/swp391/com/backend/pojo/roles/Doctor.java
@@ -1,9 +1,12 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.roles;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import swp391.com.backend.pojo.appointments.Appointment;
+import swp391.com.backend.pojo.appointments.RatingFeedback;
+import swp391.com.backend.pojo.appointments.Schedule;
 
 import java.util.List;
