+++ backend/src/main/java/swp391/com/backend/pojo/roles/Staff.java
@@ -1,15 +1,18 @@
-package swp391.com.backend.pojo;
+package swp391.com.backend.pojo.roles;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import swp391.com.backend.pojo.order.Order;
+
+import java.util.List;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
-public class Admin {
+public class Staff {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
@@ -17,5 +20,9 @@ public class Admin {
     @OneToOne
     @JoinColumn(name = "account_id", unique = true)
     private Account account;
-}
 
+    private String specialization;
+
+    @OneToMany(mappedBy = "staff")
+    private List<Order> orders;
+}
